Add files via upload
Components of a disconnected graph
An adjacency matrix for a non-connected graph will have multiple components (disconnected subgraphs) within the same matrix. Each component will have its own set of vertices and edges. Here's an example of an adjacency matrix for a non-connected graph with two components:

Let's assume we have a graph with six vertices (labeled 0 to 5) and two disconnected components: one with vertices 0, 1, and 2, and the other with vertices 3, 4, and 5. There are no edges between the two components.

The adjacency matrix for this non-connected graph would look like this:
  0 1 2 3 4 5
0 0 1 1 0 0 0
1 1 0 1 0 0 0
2 1 1 0 0 0 0
3 0 0 0 0 1 1
4 0 0 0 1 0 1
5 0 0 0 1 1 0